Refactor REPL logic from main.rs into src/tools/repl.rs
This commit extracts the REPL implementation from the main binary file
into a dedicated module in `src/tools/repl.rs`. This reduces the
complexity of `src/main.rs` (the "God Object" smell) and improves
separation of concerns.

The `run_repl` function is now exposed publicly from `glossa::tools::repl`
and called by `main`. All associated structs (`ReplContext`, `ReplOutput`)
and constants are encapsulated within the new module.

No logic changes were made; this is a pure refactor. All tests passed.
