+++ src/tools/mod.rs
@@ -1,1 +1,2 @@
 pub mod highlight;
+pub mod repl;
